requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
yfinance==0.2.44
feedparser==6.0.11
anthropic==0.40.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
httpx==0.27.2
